Start work on terminal controls refactoring
