+++ Orrery Combat Framework - Heart Module/Data/Scripts/HeartModule/Weapons/Setup/SorterTurretTerminalControls.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Heart_Module.Data.Scripts.HeartModule.Weapons.Setup
+{
+    internal class SorterTurretTerminalControls
+    {
+        // TODO
+    }
+}
